object_detection/requirements_macbook.txt
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
imutils==0.5.2
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyWavelets==1.0.1
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
toolz==0.9.0
